- Updated build.gradle.kts - Moved failing tests to jvm main, because its some problems is some files present in commonTest sourceSet - Removed deprecated parameters from gradle.properties
