+++ gradle.properties
@@ -5,10 +5,8 @@ android.useAndroidX=true
 org.gradle.parallel=true
 org.jetbrains.compose.experimental.jscanvas.enabled=true
 org.jetbrains.compose.experimental.uikit.enabled=true
-kotlin.native.enableDependencyPropagation=false
-kotlin.mpp.enableGranularSourceSetsMetadata=true
-kotlin.native.binary.memoryModel=experimental
 kotlin.mpp.enableCInteropCommonization=true
 kotlin.native.cocoapods.generate.wrapper=true
 kotlin.native.cacheKind=none
 kotlin.native.useEmbeddableCompilerJar=true
+kotlin.mpp.androidSourceSetLayoutVersion=2
